feat(dust): new action to reply to existing conversation
