+++ packages/pieces/community/dust/src/index.ts
@@ -4,6 +4,7 @@ import {
   Property,
 } from '@activepieces/pieces-framework';
 import { createConversation } from './lib/actions/create-conversation';
+import { replyToConversation } from './lib/actions/reply-to-conversation';
 
 export const dustAuth = PieceAuth.CustomAuth({
   description: 'Dust authentication requires an API key.',
@@ -32,6 +33,6 @@ export const dust = createPiece({
   minimumSupportedRelease: '0.9.0',
   logoUrl: 'https://cdn.activepieces.com/pieces/dust.png',
   authors: ['AdamSelene'],
-  actions: [createConversation],
+  actions: [createConversation, replyToConversation],
   triggers: [],
 });
